# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "remex-server",
    "description": "",
    "contact": {
      "name": "Zion Koyl",
      "email": "hiszd1@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/clients/new": {
      "post": {
        "tags": [
          "handlers::clients"
        ],
        "operationId": "create_client",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Client created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{id}": {
      "get": {
        "tags": [
          "handlers::clients"
        ],
        "operationId": "get_client_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientWithGroups"
                }
              }
            }
          },
          "404": {
            "description": "Client not found"
          }
        }
      }
    },
...
    "/groups/new": {
      "post": {
        "tags": [
          "handlers::groups"
        ],
        "operationId": "create_group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}": {
      "get": {
        "tags": [
          "handlers::groups"
        ],
        "operationId": "get_group_by_id",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "404": {
            "description": "Group not found"
          }
        }
      }
    },
    "/groups/{group_id}/clients": {
      "get": {
        "tags": [
          "handlers::groups"
        ],
        "operationId": "get_group_clients",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clients in group found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientSRV"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group not found"
          }
        }
      }
    },
    "/groups/{group_id}/jobs": {
      "get": {
        "tags": [
          "handlers::groups"
        ],
        "operationId": "get_group_jobs",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs in group found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobSRV"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group not found"
          }
        }
      }
    },
    "/groups/{group_id}/jobs/{job_id}/status": {
      "get": {
        "tags": [
          "handlers::groups"
        ],
        "operationId": "get_group_job_status_handler",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group job status found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupJobStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group or job not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Client": {
        "type": "object",
        "required": [
          "id",
          "secret",
          "client_name",
          "hardware_hash",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "client_name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "hardware_hash": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ClientSRV": {
        "type": "object",
        "required": [
          "id",
          "secret",
          "client_name",
          "hardware_hash",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "client_name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "hardware_hash": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClientStatusSummary": {
        "type": "object",
        "required": [
          "client_id",
          "client_name",
          "status"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "client_name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ClientWithGroups": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ClientSRV"
          },
          {
            "type": "object",
            "required": [
              "group_ids"
            ],
            "properties": {
              "group_ids": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "CreateClientForm": {
        "type": "object",
        "required": [
          "client_name",
          "hardware_hash"
        ],
        "properties": {
          "client_name": {
            "type": "string"
          },
          "hardware_hash": {
            "type": "string"
          }
        }
      },
      "CreateGroupForm": {
        "type": "object",
        "required": [
          "group_name"
        ],
        "properties": {
          "group_name": {
            "type": "string"
          }
        }
      },
...
      "Group": {
        "type": "object",
        "required": [
          "id",
          "group_name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "group_name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GroupJobStatusResponse": {
        "type": "object",
        "required": [
          "group_id",
          "job_id",
          "status",
          "client_statuses",
          "total_clients",
          "completed_clients",
          "failed_clients",
          "running_clients"
        ],
        "properties": {
          "client_statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientStatusSummary"
            }
          },
          "completed_clients": {
            "type": "integer",
            "format": "int64"
          },
          "failed_clients": {
            "type": "integer",
            "format": "int64"
          },
          "group_id": {
            "type": "string"
          },
          "job_id": {
            "type": "string"
          },
          "running_clients": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "total_clients": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "JobSRV": {
        "type": "object",
        "required": [
          "id",
          "job_name",
          "job_type",
          "job_status",
          "job_shell",
          "job_command",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "job_command": {
            "type": "string"
          },
          "job_name": {
            "type": "string"
          },
          "job_shell": {
            "type": "string"
          },
          "job_status": {
            "type": "string"
          },
          "job_type": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}